language clarification
I stumbled across this project, and am really enjoying reading about it. I am guessing that the documentation author isn't a native English speaker since some of the sentences are hard to read, or at least unidiomatic. So I thought I could help with that, feel free to disregard this pull request if it isn't appreciated. Although I think it would be cool if this project could reach more people, and I think clear documentation could help with that. Also if I misunderstand anything, and make the documentation faulty, then please give a heads up! :)